requirements.txt
requests==2.31.0
Pillow==10.2.0
beautifulsoup4==4.12.3
